Speed up mouse wheel scrolling
Previously, mouse wheel scrolling would send individual 1-line
scroll messages to itself. But when a mouse wheel event crosses
the threshold, it is common to scroll 3 lines.

This alone made scrolling feel slow, since it moved in 3 discrete
chunks.

But it also would add messages to the back of the queue again.

The solution is to process the mouse wheel right then and to
scroll by the correct amount.

Fixes #7