Add Buffer source and offset
Previously, the Buffer was assumed to be part of the current file.
This idea meant an implicit coupling between the Buffer and the
Workspace.

This adds a file source and offset to the Buffer class. This
removes that implicit coupling as well as makes way for future
changes such as using a shadow file.

Contributes to #3 and #29