Use the BufferChain
Previously, the Workspace would hold all of the file's contents.
This is now delegated out to the BufferChain. As the name
suggests, it holds a chain of buffers containinig all of
workspace's data. This means a file can be read in multiple chunks
or a new line is inserted in the middle.

Contributes to #3 and #29