Use common shared library
Currently, the main executable and the unit tests need to both
compiler the code being used/tested. If there was complete test
coverage, this would double the build time.

This commit adds a static library which gets used by both the
executable and unit tests. Almost all of the code moves from the
executable to this new static library.

Fixes #40