Support multiple buffers
Previously, although there was a BufferChain class, only the first
buffer in the chain would be populated with the entire file. Other
code needed to be updated to allow multiple buffers.

This commit fixes the code which relied on using only one buffer
in the buffer chain and allows true multiple-buffer support.

Closes #3