Add buffer backing
Currently, a workspace can be split across multiple buffers. This
is good for allowing files to be loaded in chunks (potentially
lazily). But it is also good for inserting a buffer in the middle
when the user types.

However, all buffers were previously assumed to be loaded from a
file. This becomes a problem if we decide to lazy load. Maybe that
chunk will exist but hasn't been loaded yet.

Additionally, we needed a way to indicate whether a buffer can be
purged to save memory or if it is in memory from what the user
typed/pasted.

This commit adds buffer backing to indicate what the buffer's
contents are from.

Contributes to #4 and #29