Use Consolas as the default font
Previously, maxHex used the system's default monospace font.
This font is quite ugly. It appears to be a bitmap font with no
antialiasing / ClearType.

This commit defaults to using the Consolas font which is available
in Windows Vista and later. If it cannot create that font, it falls
back to the system's default monospace font.

Closes #5