Fix window width calculation
Previously, the window width was calculated based off character
width only. It added padding for the vertical scroll bar.

This meant it was guestimating how many characters would make up
one scroll bar's width. This was a silly thing to do.

This commit gets the actual scroll bar width and sets the window
width to the correct amount.