Reduce over-raster
Previously, we would raster from the first visible line to the end of
the file. This includes if only the last line needs to be repainted.

Since the number of visible lines is likely to be relatively small, the
repaint of visible-but-not-invalidated is likely not critical. But it is
still wasteful.

The more important issue is that we would raster all the way to the
final line. So long files would do a lot of unnecessary raster.

This commit only rasters the lines that need to be rastered and closes #8.