Don't create separate library
Previously, maxHex was a library. A separate project was an executable
that depended on the library. The intent was to eventually have a
console executable and a GUI executable.

However, that intent can still be met without needing a library. The
code can be built twice (one for each executable).

Perhaps building twice will increase build times enough to want a common
library again. But let's cross that bridge when we get there.